refactor: change caregiverId type from Long to Integer across service, repository, and controller

+++ backend/caregiver-service/src/main/java/com/findcare/caregiver/repository/CaregiverRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
  * caregiver-specific operations.
  */
 @Repository
-public interface CaregiverRepository extends JpaRepository<Caregiver, Long> {
+public interface CaregiverRepository extends JpaRepository<Caregiver, Integer> {
 
     /**
      * Finds a caregiver by their associated account ID.
